LC-concatenate-non-zero-digits-and-multiply-by-sum-i [Runtime: N/A; Memory: N/A]

+++ math/concatenate-non-zero-digits-and-multiply-by-sum-i.cpp
@@ -3,9 +3,9 @@ public:
     long long sumAndMultiply(int n) {
        if(n==0) return 0;
 
-      int sum = 0;
-      int ans =0;
-      int i = 1;
+      long long  sum = 0;
+      long long ans =0;
+      long long i = 1;
       while(n!=0){
         int rem = n%10;
         if(rem != 0) {
